Added props and override spread attributes
