+++ packages/studio-ui-codegen-react/lib/react-component-renderer.ts
@@ -3,7 +3,7 @@ import {
   StudioComponentProperties,
 } from "@amzn/amplify-ui-codegen-schema";
 import { ComponentRendererBase } from "@amzn/studio-ui-codegen";
-import { factory, JsxAttribute, JsxElement } from "typescript";
+import { factory, JsxAttribute, JsxAttributeLike, JsxElement, JsxOpeningElement, NodeFactory } from "typescript";
 
 import { ImportCollection } from "./import-collection";
 
@@ -44,5 +44,51 @@ export abstract class ReactComponentRenderer<
     }
 
     return propsArray;
+  }
+
+  protected renderOpeningElement(
+    factory: NodeFactory,
+    props: StudioComponentProperties,
+    tagName: string
+  ): JsxOpeningElement {
+    const propsArray: JsxAttribute[] = [];
+    for (let propKey of Object.keys(props)) {
+      const currentProp = props[propKey];
+
+      if (currentProp.value) {
+        const attr = factory.createJsxAttribute(
+          factory.createIdentifier(propKey),
+          factory.createStringLiteral(currentProp.value)
+        );
+
+        propsArray.push(attr);
+      }
+    }
+
+    this.addPropsSpreadAttributes(factory, propsArray, tagName);
+
+    return factory.createJsxOpeningElement(
+      factory.createIdentifier(tagName),
+      undefined,
+      factory.createJsxAttributes(propsArray)
+    );
+  }
+
+  private addPropsSpreadAttributes(
+    factory: NodeFactory,
+    attributes: JsxAttributeLike[],
+    tagName: string
+  ) {
+    const propsAttr = factory.createJsxSpreadAttribute(factory.createIdentifier("props"));
+    attributes.push(propsAttr);
+
+    const overrideAttr = factory.createJsxSpreadAttribute(
+      factory.createCallExpression(
+        factory.createIdentifier("getOverrideProps"), 
+        undefined, 
+        [factory.createIdentifier("overrides"), factory.createStringLiteral(tagName)]
+      )
+    );
+    attributes.push(overrideAttr);
   }
 }
